backend/requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2==2.8.6
gunicorn==20.1.0